// apps/admin/src/widgets/ShopOwnerLayout/types/sidebarListEnum.ts
export enum SidebarListEnum {
    Statistic = 'Statistic',
    SupplierStatistic = 'SupplierStatistic',
    OrderTable = 'OrderTable',
    OrderList = 'OrderList',
    DeliveryDistrict = 'DeliveryDistrict',
    PickupPoint = 'PickupPoint',
    Suppliers = 'Suppliers',
    SupplierOrderTable = 'SupplierOrderTable',
    CreateProduct = 'CreateProduct',
    CreateTelegramBroadcast = 'CreateTelegramBroadcast',
    ShopSettings = 'ShopSettings',
    SupplierProducts = 'SupplierProducts'
}